fix: add missing exports in blog-service.ts
Correct missing exports in blog-service.ts file.

#VERCEL_SKIP

+++ scripts/create-view-increment-function.sql
@@ -1,9 +1,9 @@
--- Create a function to safely increment post views
-CREATE OR REPLACE FUNCTION increment_post_views(post_id BIGINT)
+-- Create function to increment post views
+CREATE OR REPLACE FUNCTION increment_post_views(post_id INTEGER)
 RETURNS VOID AS $$
 BEGIN
-  UPDATE blog_posts 
-  SET views = views + 1 
-  WHERE id = post_id;
+    UPDATE blog_posts 
+    SET views = views + 1 
+    WHERE id = post_id;
 END;
 $$ LANGUAGE plpgsql;
